Intermediate refactoring for AuthorityType
