+++ src/client/Microsoft.Identity.Client/Internal/Requests/UsernamePasswordRequest.cs
@@ -46,8 +46,7 @@ namespace Microsoft.Identity.Client.Internal.Requests
 
         private async Task<UserAssertion> FetchAssertionFromWsTrustAsync()
         {
-            if (AuthenticationRequestParameters.AuthorityInfo.AuthorityType == AuthorityType.Adfs ||
-                AuthenticationRequestParameters.AuthorityInfo.AuthorityType == AuthorityType.B2C)
+            if (!AuthenticationRequestParameters.AuthorityInfo.IsUserAssertionSupported)
             {
                 return null;
             }
